Use deepcell Application class
